--- FilmApps/app/src/main/java/com/example/filmapps/Result.kt
@@ -1,6 +1,0 @@
-package com.example.filmapps
-
-sealed class Result {
-    object Success : Result()
-    class   Error(val message: String) : Result()
-}
