+++ FilmApps/app/src/main/java/com/example/filmapps/data/UserDataBase.kt
@@ -3,7 +3,7 @@ package com.example.filmapps.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.filmapps.data.model.UserData
+import com.example.filmapps.feature.regAndAuth.data.model.UserData
 
 @Database(entities = [UserData::class], version = 1, exportSchema = false)
 abstract class UserDataBase : RoomDatabase() {
