--- FilmApps/app/src/main/java/com/example/filmapps/data/model/UserData.kt
@@ -1,16 +1,0 @@
-package com.example.filmapps.data.model
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-data class UserData (
-    @ColumnInfo
-    var login: String,
-    @ColumnInfo
-    var pass: String,
-) {
-    @PrimaryKey(autoGenerate = true)
-    var id: Int? = null
-}
