--- FilmApps/app/src/main/java/com/example/filmapps/data/repositories/UserRepositories.kt
@@ -1,10 +1,0 @@
-package com.example.filmapps.data.repositories
-
-import com.example.filmapps.domain.model.UserDataParam
-
-interface UserRepositories {
-
-    fun saveLoginAndPassword(userParam: UserDataParam)
-
-    fun getInfoAboutRegisteredOrNot(userParam: UserDataParam) : Boolean
-}
