--- FilmApps/app/src/main/java/com/example/filmapps/domain/di/module/ViewModelKey.kt
@@ -1,14 +1,0 @@
-package com.example.filmapps.domain.di.module
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-
-@MustBeDocumented
-@Target(AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER)
-@Retention(AnnotationRetention.RUNTIME)
-@MapKey
-annotation class ViewModelKey(val value: KClass<out ViewModel>)
