--- FilmApps/app/src/main/java/com/example/filmapps/domain/di/module/ViewModelModule.kt
@@ -1,14 +1,0 @@
-package com.example.filmapps.domain.di.module
-
-import androidx.lifecycle.ViewModelProvider
-import com.example.filmapps.presentation.viewModel.ViewModelFactory
-import dagger.Binds
-import dagger.Module
-
-
-@Module
-interface ViewModelModule {
-
-    @Binds
-    fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
-}
