--- FilmApps/app/src/main/java/com/example/filmapps/domain/model/UserDataParam.kt
@@ -1,3 +1,0 @@
-package com.example.filmapps.domain.model
-
-class UserDataParam(val login: String, val pass: String)
