--- FilmApps/app/src/main/java/com/example/filmapps/domain/useCase/SaveUserDataUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.filmapps.domain.useCase
-
-import com.example.filmapps.domain.model.UserDataParam
-
-interface SaveUserDataUseCase {
-    fun execute(param: UserDataParam): Boolean
-}
