--- FilmApps/app/src/main/java/com/example/filmapps/domain/useCase/SaveUserDataUseCaseImpl.kt
@@ -1,18 +1,0 @@
-package com.example.filmapps.domain.useCase
-
-import com.example.filmapps.data.repositories.UserRepositories
-import com.example.filmapps.domain.model.UserDataParam
-import javax.inject.Inject
-
-internal class SaveUserDataUseCaseImpl @Inject constructor(
-    private val userRepository: UserRepositories
-) : SaveUserDataUseCase {
-    override fun execute(param: UserDataParam): Boolean {
-        if (param.login.isEmpty() || param.pass.isEmpty())
-            return false
-        else {
-            userRepository.saveLoginAndPassword(userParam = param)
-            return true
-        }
-    }
-}
