--- FilmApps/app/src/main/java/com/example/filmapps/domain/useCase/getInfoAboutRegisterUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.filmapps.domain.UseCase
-
-import com.example.filmapps.domain.model.UserDataParam
-
-interface GetInfoAboutRegisterUseCase {
-    fun execute(param: UserDataParam): Boolean
-}
