+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/Result.kt
@@ -1,0 +1,6 @@
+package com.example.filmapps.feature.regAndAuth
+
+sealed class Result {
+    object Success : Result()
+    class   Error(val message: String) : Result()
+}
