+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/authorization/domain/di/scopes/AuthorizationScope.kt
@@ -1,0 +1,6 @@
+package com.example.filmapps.feature.regAndAuth.authorization.domain.di.scopes
+
+import javax.inject.Scope
+
+@Scope
+annotation class AuthorizationScope
