+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/authorization/domain/useCase/GetInfoAboutRegisterUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.filmapps.feature.regAndAuth.authorization.domain.useCase
+
+import com.example.filmapps.feature.regAndAuth.domain.model.UserDataParam
+
+interface GetInfoAboutRegisterUseCase {
+    fun execute(param: UserDataParam): Boolean
+}
