+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/data/repositories/UserRepositories.kt
@@ -1,0 +1,10 @@
+package com.example.filmapps.feature.regAndAuth.data.repositories
+
+import com.example.filmapps.feature.regAndAuth.domain.model.UserDataParam
+
+interface UserRepositories {
+
+    fun saveLoginAndPassword(userParam: UserDataParam)
+
+    fun getInfoAboutRegisteredOrNot(userParam: UserDataParam) : Boolean
+}
