+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/domain/di/module/ViewModelModule.kt
@@ -1,0 +1,14 @@
+package com.example.filmapps.feature.regAndAuth.domain.di.module
+
+import androidx.lifecycle.ViewModelProvider
+import com.example.filmapps.feature.regAndAuth.presentation.viewModel.ViewModelFactory
+import dagger.Binds
+import dagger.Module
+
+
+@Module
+interface ViewModelModule {
+
+    @Binds
+    fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
+}
