+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/domain/model/UserDataParam.kt
@@ -1,0 +1,3 @@
+package com.example.filmapps.feature.regAndAuth.domain.model
+
+class UserDataParam(val login: String, val pass: String)
