+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/presentation/viewModel/ViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.example.filmapps.feature.regAndAuth.presentation.viewModel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import javax.inject.Inject
+import javax.inject.Provider
+
+class ViewModelFactory @Inject constructor(private val viewModels: MutableMap<Class<out ViewModel>, Provider<ViewModel>>) : ViewModelProvider.Factory {
+
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        val viewModelProvider = viewModels[modelClass]
+            ?: throw IllegalArgumentException("model class $modelClass not found")
+        return viewModelProvider.get() as T
+    }
+}
