+++ FilmApps/app/src/main/java/com/example/filmapps/feature/regAndAuth/registrations/domain/useCase/SaveUserDataUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.filmapps.feature.regAndAuth.registrations.domain.useCase
+
+import com.example.filmapps.feature.regAndAuth.domain.model.UserDataParam
+
+interface SaveUserDataUseCase {
+    fun execute(param: UserDataParam): Boolean
+}
