+++ FilmApps/app/src/main/res/layout/activity_main.xml
@@ -23,7 +23,8 @@
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/layoutCont"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior"/>
 
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
