+++ FilmApps/app/src/main/res/layout/fragment_list_film2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Presentation.UI.ListFilmFragment">
+    tools:context=".presentation.ui.ListFilmFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
